requirements.txt
praw>=7.7.1
pyyaml>=6.0
